requirements.txt
numpy==1.23.5
Pillow==9.2.0
Pillow==9.5.0
Pillow==9.5.0
polyglot==15.10.03
pytesseract==0.3.10
python_Levenshtein==0.21.0
tensorflow==2.10.1
tensorflow_intel==2.12.0
Levenshtein~=0.21.0